Refactor BaseModel
